Intento #10

+++ src/actions/address/delete-user-address.ts
@@ -7,10 +7,22 @@ import prisma from '@/lib/prisma';
 export const deleteUserAddress = async( userId: string ) => {
 
   try {
+    if (!userId){
+      return {
+        ok: false,
+        message: 'Debe proporcionar un id de usuario'
+      }
+    }
 
     const deleted = await prisma.userAddress.delete({
       where: { userId }
     });
+    if(!deleted){
+      return {
+        ok: false,
+        message: 'No se encontró la dirección para el usuario'
+      }
+    }
 
     return { ok: true };
